pyproject.toml
[project]
name = "briefbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiogram>=3.26.0",
    "aiohttp>=3.12.15",
    "anthropic>=0.54.0",
    "beautifulsoup4>=4.13.5",
    "groq>=0.31.0",
    "httpx>=0.28.1",
    "openai==2.26.0",
    "pypdf>=6.0.0",
    "python-dotenv>=1.2.2",
    "supabase>=2.28.0",
    "youtube-transcript-api>=0.6.2",
]
